Cleanup JobSupport slightly
Before this change, there were some extra classes/interfaces that
were immediately wrapped in other classes and so serve no real
purpose. These classes are removed, slightly reducing the number
of allocations in some cases and simplifying the code.
