+++ kotlinx-coroutines-core/common/src/CompletionHandler.common.kt
@@ -25,47 +25,3 @@ package kotlinx.coroutines
  */
 // TODO: deprecate. This doesn't seem better than a simple function type.
 public typealias CompletionHandler = (cause: Throwable?) -> Unit
-
-/**
- * Essentially the same as just a function from `Throwable?` to `Unit`.
- * The only thing implementors can do is call [invoke].
- * The reason this abstraction exists is to allow providing a readable [toString] in the list of completion handlers
- * as seen from the debugger.
- * Use [UserSupplied] to create an instance from a lambda.
- * We can't avoid defining a separate type, because on JS, you can't inherit from a function type.
- *
- * @see CancelHandler for a very similar interface, but used for handling cancellation and not completion.
- */
-internal interface InternalCompletionHandler {
-    /**
-     * Signals completion.
-     *
-     * This function:
-     * - Does not throw any exceptions.
-     *   For [Job] instances that are coroutines, exceptions thrown by this function will be caught, wrapped into
-     *   [CompletionHandlerException], and passed to [handleCoroutineException], but for those that are not coroutines,
-     *   they will just be rethrown, potentially crashing unrelated code.
-     * - Is fast, non-blocking, and thread-safe.
-     * - Can be invoked concurrently with the surrounding code.
-     * - Can be invoked from any context.
-     *
-     * The meaning of `cause` that is passed to the handler is:
-     * - It is `null` if the job has completed normally.
-     * - It is an instance of [CancellationException] if the job was cancelled _normally_.
-     *   **It should not be treated as an error**. In particular, it should not be reported to error logs.
-     * - Otherwise, the job had _failed_.
-     */
-    fun invoke(cause: Throwable?)
-
-    /**
-     * A lambda passed from outside the coroutine machinery.
-     *
-     * See the requirements for [InternalCompletionHandler.invoke] when implementing this function.
-     */
-    class UserSupplied(private val handler: (cause: Throwable?) -> Unit) : InternalCompletionHandler {
-        /** @suppress */
-        override fun invoke(cause: Throwable?) { handler(cause) }
-
-        override fun toString() = "InternalCompletionHandler.UserSupplied[${handler.classSimpleName}@$hexAddress]"
-    }
-}
