+++ kotlinx-coroutines-core/common/src/Exceptions.common.kt
@@ -1,7 +1,7 @@
 package kotlinx.coroutines
 
 /**
- * This exception gets thrown if an exception is caught while processing [InternalCompletionHandler] invocation for [Job].
+ * This exception gets thrown if an exception is caught while processing [CompletionHandler] invocation for [Job].
  *
  * @suppress **This an internal API and should not be used from general code.**
  */
